Move junction restrictions structures out of API
