+++ TLM/TMPE.API/TMPE.API.csproj
@@ -1,4 +1,4 @@
-<?xml version="1.0" encoding="utf-8"?>
+﻿<?xml version="1.0" encoding="utf-8"?>
 <Project ToolsVersion="14.0" DefaultTargets="Build" xmlns="http://schemas.microsoft.com/developer/msbuild/2003">
   <Import Project="$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props" Condition="Exists('$(MSBuildExtensionsPath)\$(MSBuildToolsVersion)\Microsoft.Common.props')" />
   <PropertyGroup>
@@ -130,8 +130,6 @@
     <Compile Include="Traffic\Data\PathUnitQueueItem.cs" />
     <Compile Include="Traffic\Data\PrioritySegment.cs" />
     <Compile Include="Traffic\Data\SegmentDirTrafficData.cs" />
-    <Compile Include="Traffic\Data\SegmentEndFlags.cs" />
-    <Compile Include="Traffic\Data\SegmentFlags.cs" />
     <Compile Include="Traffic\Data\ExtVehicle.cs" />
     <Compile Include="Traffic\Data\SpeedValue.cs" />
     <Compile Include="Traffic\Data\TurnOnRedSegments.cs" />
